### 2021-03-17 #3120 @whedon
[PRE REVIEW]: TRUNAJOD: A text complexity library to enhance natural language processing
**Submitting author:** @dpalmasan (<a href="http://orcid.org/0000-0003-1540-7164">Diego Palma</a>)
**Repository:** <a href="https://github.com/dpalmasan/TRUNAJOD2.0" target ="_blank">https://github.com/dpalmasan/TRUNAJOD2.0</a>
**Version:** 0.1.0.0
