### 2021-03-17 #3120 @dpalmasan
These reviewers might be suitable:

ankur-gupta, shanamatthews, mbdemoraes/moafs, colbrydi, apiad

El mié, 17 mar 2021 a las 8:40, whedon ***@***.***>)
escribió:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - None
>
> MISSING DOIs
>
> - 10.3758/s13428-015-0651-7 may be a valid DOI for title: The tool for the automatic analysis of text cohesion (TAACO): Automatic assessment of local, global, and text cohesion
> - 10.1080/01638539809545029 may be a valid DOI for title: The measurement of textual coherence with latent semantic analysis
> - 10.1080/00220973.1994.9943835 may be a valid DOI for title: Computer grading of student prose, using modern concepts and software
> - 10.3758/s13428-015-0684-y may be a valid DOI for title: Spanish norms for affective and lexico-semantic variables for 1,400 words
>
> INVALID DOIs
>
> - None
>
