### 2021-03-28 8710035
Stanza option added to TTR, Entity Grid
