+++ stanza_test.py
@@ -1,0 +1,43 @@
+from src.TRUNAJOD.entity_grid import EntityGrid
+from src.TRUNAJOD.ttr import lexical_diversity_mtld, one_side_lexical_diversity_mtld
+import spacy
+import stanza
+"""
+MUST CHANGE TTR IMPORT TO WORK
+"""
+# Load spaCy model
+nlp = spacy.load("es_core_news_sm")
+
+# Load stanza model
+nlp_s = stanza.Pipeline('es', use_gpu=False)
+
+# Example
+example_text = (
+    "El espectáculo del cielo nocturno cautiva la mirada y suscita preguntas"
+    "sobre el universo, su origen y su funcionamiento. No es sorprendente que "
+    "todas las civilizaciones y culturas hayan formado sus propias "
+    "cosmologías. Unas relatan, por ejemplo, que el universo ha"
+    "sido siempre tal como es, con ciclos que inmutablemente se repiten; "
+    "otras explican que este universo ha tenido un principio, "
+    "que ha aparecido por obra creadora de una divinidad."
+)
+
+# Create Doc
+doc = nlp(example_text)
+doc_s = nlp_s(example_text)
+
+# TTR Check - change TTR import to test
+print("spacy result: ", lexical_diversity_mtld(doc))
+# or
+# print("spacy result: ", lexical_diversity_mtld(doc, model="spacy"))
+print("stanza result: ", lexical_diversity_mtld(doc_s, model="stanza"))
+
+# Entity Grid Check
+egrid = EntityGrid(doc)
+egrid_s = EntityGrid(doc_s, model="stanza")
+
+print("spacy Entity grid:")
+print(egrid.get_egrid())
+
+print("stanza Entity grid:")
+print(egrid_s.get_egrid())
