### 2021-03-29 a911578
test conflict resolve and enum implementation
