+++ src/TRUNAJOD/ttr.py
@@ -8,7 +8,7 @@ the text this measurement is 1, and if there is infinite repetition, it will
 tend to 0. This measurement is not recommended if analyzing texts of different
 lengths, as when the number of tokens increases, the TTR tends flatten.
 """
-from TRUNAJOD.utils import is_word
+from TRUNAJOD.utils import SupportedModels,is_word
 
 # dev import
 # from src.TRUNAJOD.utils import is_word
@@ -25,7 +25,7 @@ def type_token_ratio(word_list):
     return len(set(word_list)) / len(word_list)
 
 
-def lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
+def lexical_diversity_mtld(doc, model_name="spacy", ttr_segment=0.72):
     """Compute MTLD lexical diversity in a bi-directional fashion.
 
     :param doc: Processed text
@@ -33,12 +33,15 @@ def lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
     :return: Bi-directional lexical diversity MTLD
     :rtype: float
     """
+    # check model
+    model = SupportedModels(model_name)
+
     word_list = []
-    if model == "spacy":
+    if model == SupportedModels.SPACY:
         for token in doc:
             if is_word(token.pos_):
                 word_list.append(token.lemma_)
-    elif model == "stanza":
+    elif model == SupportedModels.STANZA:
         for sent in doc.sentences:
             for word in sent.words:
                 if is_word(word.upos):
@@ -47,7 +50,7 @@ def lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
             one_side_lexical_diversity_mtld(word_list[::-1], model, ttr_segment)) / 2
 
 
-def one_side_lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
+def one_side_lexical_diversity_mtld(doc, model_name="spacy", ttr_segment=0.72):
     """Lexical diversity per MTLD.
 
     :param doc: Tokenized text
@@ -62,7 +65,10 @@ def one_side_lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
     non_ttr_segment = 1 - ttr_segment
     word_list = []
 
-    if model == "spacy" or type(doc) == list:
+    # check model
+    model = SupportedModels(model_name)
+
+    if model == SupportedModels.SPACY or type(doc) == list:
         for token in doc:
             word_list.append(token.lower())
             total_words += 1
@@ -70,7 +76,7 @@ def one_side_lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
             if ttr < ttr_segment:
                 word_list = []
                 factor += 1
-    elif model == "stanza":
+    elif model == SupportedModels.STANZA:
         if type(doc) != list:
             for sent in doc.sentences:
                 for word in sent.words:
@@ -85,5 +91,4 @@ def one_side_lexical_diversity_mtld(doc, model="spacy", ttr_segment=0.72):
         factor += 1 - (
             type_token_ratio(word_list) - ttr_segment) / non_ttr_segment
         total_words += 1
-
     return total_words / factor
