### 2021-04-04 b5e102f
Added type hints to functions in discourse_markers.py
