### 2021-04-04 bda5f86
Added "pip install spacy" in .travis.yml, under install
