### 2021-04-05 2642c5f
Removed spacy install in travis.yml, Added spacy import in discourse_markers, Added spacy dependency in tox.ini
