+++ tox.ini
@@ -10,4 +10,5 @@ deps =
     coverage
     pytest-cov
     mock
+    spacy
 commands = pytest
