### 2021-04-05 77c50fa
Changed files to use Python 3.7 for tests and ran pre-commit
