+++ .travis.yml
@@ -1,6 +1,6 @@
 language: python
 python:
-  - 3.6
+  - 3.7
 before_install:
   - python --version
   - pip install -U pip
