### 2021-04-05 ff218e2
Changed files to use Python 3.7 for tests
