+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py36
+envlist = py37
 
 [pytest]
 norecursedirs = docs *.egg-info .git appdir .tox
@@ -11,4 +11,5 @@ deps =
     pytest-cov
     mock
     spacy
+
 commands = pytest
