### 2021-04-06 e69a3cf
Add style checks to travis.
* Update travis to run style checks
* Add contributing docs
* Update flake8 failing scripts
