+++ src/TRUNAJOD/semantic_measures.py
@@ -79,9 +79,8 @@ def overlap(lemma_list_group, synset_dict):
     """
     if len(lemma_list_group) < 2:  # pragma: no cover
         raise RuntimeError(
-            "At least two sentences should be provided, you provided {}".format(
-                lemma_list_group
-            )
+            "At least two sentences should be "
+            "provided, you provided {}".format(lemma_list_group)
         )
 
     N = len(lemma_list_group)
