+++ .travis.yml
@@ -1,12 +1,17 @@
 language: python
 python:
-  - 3.6
+  - 3.7
 before_install:
   - python --version
   - pip install -U pip
   - pip install -U pytest
   - pip install codecov
   - pip install -U pytest-cov
+  - pip install flake8
+  - pip install black
 install:
   - pip install tox-travis
-script: tox
+scripts:
+  - black src
+  - flake8 src
+  - tox
