+++ setup.cfg
@@ -1,3 +1,8 @@
 [metadata]
 description-file = README.md
 long_description_content_type = text/markdown
+
+[flake8]
+extend-ignore =
+    # See https://github.com/PyCQA/pycodestyle/issues/373
+    E203,
