+++ .pre-commit-hooks.yaml
@@ -5,19 +5,17 @@ repos:
     hooks:
     -    id: black
          stages: [commit]
-         language_version: python3.6
+         language_version: python3.7
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.5.0
     hooks:
     -   id: flake8
         stages: [commit]
-        id: check-docstring-first
-        stages: [commit]
 -   repo: https://github.com/asottile/reorder_python_imports
     rev: v2.1.0
     hooks:
     -   id: reorder-python-imports
-        language_version: python3.6
+        language_version: python3.7
 -   repo: git://github.com/chewse/pre-commit-mirrors-pydocstyle
     rev: v2.1.1
     hooks:
