### 2021-04-06 #3153 @whedon
[REVIEW]: TRUNAJOD: A text complexity library to enhance natural language processing
**Submitting author:** @dpalmasan (<a href="http://orcid.org/0000-0003-1540-7164">Diego Palma</a>)
**Repository:** <a href="https://github.com/dpalmasan/TRUNAJOD2.0" target ="_blank">https://github.com/dpalmasan/TRUNAJOD2.0</a>
**Version:** v0.1.2
