### 2021-04-06 #3153 @danielskatz
👋 @dpalmasan, can you work on the possibly missing DOIs that whedon suggests, noting that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to the be the first entry in a new comment.
