+++ paper.bib
@@ -8,14 +8,23 @@
 }
 
 @article{foltz1998measurement,
-  title={The measurement of textual coherence with latent semantic analysis},
-  author={Foltz, Peter W and Kintsch, Walter and Landauer, Thomas K},
-  journal={Discourse processes},
-  volume={25},
-  number={2-3},
-  pages={285--307},
-  year={1998},
-  publisher={Taylor \& Francis}
+  author = { Peter W.   Foltz  and  Walter   Kintsch  and  Thomas K   Landauer },
+  title = {The measurement of textual coherence with latent semantic analysis},
+  journal = {Discourse Processes},
+  volume = {25},
+  number = {2-3},
+  pages = {285-307},
+  year  = {1998},
+  publisher = {Routledge},
+  doi = {10.1080/01638539809545029},
+
+  URL = { 
+          https://doi.org/10.1080/01638539809545029
+      
+  },
+  eprint = { 
+          https://doi.org/10.1080/01638539809545029
+  }
 }
 
 @article{barzilay2008modeling,
@@ -45,14 +54,24 @@
 }
 
 @article{page1994computer,
-  title={Computer grading of student prose, using modern concepts and software},
-  author={Page, Ellis Batten},
-  journal={The Journal of experimental education},
-  volume={62},
-  number={2},
-  pages={127--142},
-  year={1994},
-  publisher={Taylor \& Francis}
+  author = { Ellis   Batten   Page },
+  title = {Computer Grading of Student Prose, Using Modern Concepts and Software},
+  journal = {The Journal of Experimental Education},
+  volume = {62},
+  number = {2},
+  pages = {127-142},
+  year  = {1994},
+  publisher = {Routledge},
+  doi = {10.1080/00220973.1994.9943835},
+
+  URL = { 
+          https://doi.org/10.1080/00220973.1994.9943835
+      
+  },
+  eprint = { 
+          https://doi.org/10.1080/00220973.1994.9943835
+      
+  }
 }
 
 @article{guasch2016spanish,
@@ -63,7 +82,13 @@
   number={4},
   pages={1358--1369},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.3758/s13428-015-0684-y},
+
+  URL = { 
+          https://doi.org/10.3758/s13428-015-0684-y
+      
+  },
 }
 
 @article{sidorov2012empirical,
