### 2021-04-06 01b8a2c
Fix CONTRIBUTING to use correct project name
