### 2021-04-10 704c9a0
Added docstring for some parameters. Added type hints.
