### 2021-04-11 f1d8643
Migrating from travis CI to github actions.
