+++ tests/surface_proxies_test.py
@@ -3,13 +3,13 @@ from collections import namedtuple
 
 from TRUNAJOD import surface_proxies
 
-Token = namedtuple("Token", "word pos_ lower_ lemma_")
+Token = namedtuple("Token", ["word", "pos_", "lower_", "lemma_", "tag_"])
 doc = [
-    Token("El", "", "el", "El"),
-    Token("perro", "NOUN", "perro", "perro"),
-    Token("es", "", "es", "es"),
-    Token("extraordinario", "ADJ", "extraordinario", "extraordinario"),
-    Token(".", "PUNCT", ".", "."),
+    Token("El", "", "el", "El", "Person=1"),
+    Token("perro", "NOUN", "perro", "perro", "tag"),
+    Token("es", "", "es", "es", "tag"),
+    Token("extraordinario", "ADJ", "extraordinario", "extraordinario", "tag"),
+    Token(".", "PUNCT", ".", ".", "tag"),
 ]
 
 
@@ -36,3 +36,16 @@ def test_negation_density():
 def test_noun_count():
     """Test noun_count."""
     assert surface_proxies.noun_count(doc) == 1
+
+
+def test_first_second_person_density():
+    """Test first_second_person_density."""
+    assert surface_proxies.first_second_person_count(doc) == 1
+
+
+def test_infinitive():
+    """Test infinitive."""
+    assert surface_proxies.infinitve("comiendo", {}) is None
+    assert (
+        surface_proxies.infinitve("comiendo", {"comiendo": "comer"}) == "comer"
+    )
