+++ .pre-commit-config.yaml
@@ -5,7 +5,6 @@ repos:
     hooks:
     -    id: black
          stages: [commit]
-         language_version: python3.6
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.5.0
     hooks:
@@ -17,7 +16,6 @@ repos:
     rev: v2.1.0
     hooks:
     -   id: reorder-python-imports
-        language_version: python3.6
 -   repo: git://github.com/chewse/pre-commit-mirrors-pydocstyle
     rev: 'v2.1.1'
     hooks:
