+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# Changelog
+
+
+## v0.1.1
+
+* Add `spaCy` as a required dependency (#23).
+* Fix pypi `long_description` (#23).
+* Migrate from `travis-ci` to `Github` actions (#22).
+* Add Type Hints to TTR module (#18)
+* Add Type Hints to discourse Markers (#16)
