### 2021-04-13 a54781f
Implement Yule K lexical diversity index (#27)
