+++ docs/conf.py
@@ -21,7 +21,7 @@ copyright = "2020, Diego Palma"
 author = "Diego Palma"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.1"
+release = "0.1.1"
 
 # -- General configuration ---------------------------------------------------
 
