### 2021-04-14 f8ad660
Update README description.
Most of the indices could be computed for different languages (e.g. English).
