### 2021-04-14 dpalmasan/TRUNAJOD2.0#39 @mbdemoraes
Specify Python version
The supported python version should be explicitly specified in the README.

cc https://github.com/openjournals/joss-reviews/issues/3153
