### 2021-04-14 dpalmasan/TRUNAJOD2.0#40 @mbdemoraes
Paper grammar review
There are some grammar issues on the paper. For instance, using "have" instead of "has", missing articles (such as "the" text), missing hyphen (such as grid-based), among others. I suggest using Grammarly (https://app.grammarly.com/) to assist the authors on this task.

cc https://github.com/openjournals/joss-reviews/issues/3153
