### 2021-04-14 dpalmasan/TRUNAJOD2.0#40 @dpalmasan
Hello @mbdemoraes Thanks for the detailed review! I have updated the `paper.md` file, and used Grammarly as suggested (I wasn't aware that it had a free version, it is very useful). I already committed the changes, so let me know should I change anything else in the `paper.md`, I will also regenerate the paper in the JOSS review
