### 2021-04-14 ce60643
Add python versions to README (#39)
