### 2021-04-14 dpalmasan/TRUNAJOD2.0#39 @dpalmasan
I have added to the `README` the supported `python` versions. Let me know if anything is left.
