### 2021-04-14 9725767
Added typehints following doctype that return types are all floats  TESTS PASS
