### 2021-04-14 d5bd937
Added typing timports and type hints in constructor
