+++ src/TRUNAJOD/emotions.py
@@ -14,6 +14,9 @@ sorpresa, tristeza.
 """
 from TRUNAJOD.spanish_emotion_lexicon import SPANISH_EMOTION_LEXICON
 from TRUNAJOD.utils import lemmatize
+from spacy.tokens import Doc
+from typing import Dict
+from typing import Optional
 
 
 class Emotions(object):
@@ -29,7 +32,7 @@ class Emotions(object):
     - Averaging over emotions and not the total count?
     """
 
-    def __init__(self, doc, lemmatizer=None):
+    def __init__(self, doc: Doc, lemmatizer: Optional[Dict[str, str]]=None):
         """Initialize emotions class.
 
         Average over number of tokens.
