### 2021-04-14 1be6f3f
Make spacy model installation explicit (#46)
