### 2021-04-14 dpalmasan/TRUNAJOD2.0#40 @mbdemoraes
@dpalmasan I'm glad to know that Grammarly helped! For us, non-native English speakers, English grammar can be quite challenging. I've checked and everything is ok now. I'm going to close the issue.
