+++ src/TRUNAJOD/givenness.py
@@ -5,6 +5,7 @@ Giveness is defined as the amount of given information a text exposes over
 successive constituents :cite:`hempelmann2005using`. Givenness is can be used
 as a proxy of text complexity.
 """
+from spacy.tokens import Doc
 from TRUNAJOD.utils import is_noun
 from TRUNAJOD.utils import is_pronoun
 from TRUNAJOD.utils import is_word
@@ -13,7 +14,7 @@ from TRUNAJOD.utils import is_word
 THIRD_PERSON_LABEL = "Person=3"
 
 
-def pronoun_density(doc : spacy.doc) -> float:
+def pronoun_density(doc: Doc) -> float:
     """Compute pronoun density.
 
     This is a measurement of text complexity, in the sense that a text
@@ -38,7 +39,7 @@ def pronoun_density(doc : spacy.doc) -> float:
     return float(third_person_pronouns) / word_counter
 
 
-def pronoun_noun_ratio(doc : spacy.doc) -> float:
+def pronoun_noun_ratio(doc: Doc) -> float:
     """Compute Pronoun Noun ratio.
 
     This is an approximation of text complexity/readability, since pronouns
