### 2021-04-14 665f04f
hopefully fixes type not subscriptable error
