+++ src/TRUNAJOD/syllabizer.py
@@ -42,7 +42,6 @@ The following rules are applied:
 | CC-CC          | If four consonants are joined, they are halved.            |
 +----------------+------------------------------------------------------------+
 """
-
 from typing import TypeVar
 
 STRONG_VOWELS = {"a", "á", "e", "é", "o", "ó", "í", "ú"}
@@ -70,7 +69,8 @@ RULES = [
 ]
 
 
-CharLine = TypeVar('CharLine')
+CharLine = TypeVar("CharLine")
+
 
 class CharLine(object):
     """Auxiliary object to set char types on a word.
@@ -93,7 +93,7 @@ class CharLine(object):
         self.type_line = "".join(chartype for _, chartype in charline)
 
     @staticmethod
-    def char_type(char : str) -> str:
+    def char_type(char: str) -> str:
         """Get char type (vowel, consonant, etc).
 
         This method checks a ``char`` type based on syllabization rules.
@@ -116,7 +116,7 @@ class CharLine(object):
         # c stands for consonant
         return char if char in {"x", "s"} else "c"
 
-    def find(self, finder : str) -> int:
+    def find(self, finder: str) -> int:
         """Find string occurrence in the type representation.
 
         :param finder: String to be searched
@@ -126,7 +126,7 @@ class CharLine(object):
         """
         return self.type_line.find(finder)
 
-    def split(self, pos : int, where : int) -> tuple[CharLine, CharLine]:
+    def split(self, pos: int, where: int) -> [CharLine, CharLine]:
         """Split the object into two Charline objects.
 
         :param pos: Start position of the split
@@ -141,7 +141,7 @@ class CharLine(object):
             CharLine(self.word[pos + where :]),
         )
 
-    def split_by(self, finder : str, where : int) -> tuple[CharLine, CharLine]:
+    def split_by(self, finder: str, where: int) -> [CharLine, CharLine]:
         """Split charline by `finder` occurrence on `type_char`.
 
         :param finder: Type char string
@@ -173,7 +173,7 @@ class CharLine(object):
         """
         return "<" + repr(self.word) + ":" + self.type_line + ">"
 
-    def __eq__(self, other : CharLine) -> bool:
+    def __eq__(self, other: CharLine) -> bool:
         """Equal operator implementation.
 
         :param other: CharLine to be compared to.
@@ -192,7 +192,7 @@ class Syllabizer(object):
     """
 
     @staticmethod
-    def split(chars : CharLine) -> list[CharLine]:
+    def split(chars: CharLine) -> list[CharLine]:
         """Split CharLine into syllabes.
 
         :param chars: Word to be syllabized
@@ -226,7 +226,7 @@ class Syllabizer(object):
         return [chars]
 
     @staticmethod
-    def number_of_syllables(word : str) -> int:
+    def number_of_syllables(word: str) -> int:
         """Return number of sillables of a word.
 
         :param word: Word to be processed
