### 2021-04-14 d4fa584
remove forced type to allow compatibility over all OS'
