### 2021-04-15 74dc824
Add D-estimate and update refs. (# 29)
