+++ tests/ttr_test.py
@@ -1,7 +1,27 @@
 """TRUNAJOD ttr tests."""
+import string
 from collections import namedtuple
 
+import numpy as np
+import pytest
 from TRUNAJOD import ttr
+
+
+Token = namedtuple("Token", ["lemma_", "pos_"])
+
+
+@pytest.fixture
+def test_doc():
+    """Fixture to use a doc for tests."""
+    doc = [
+        Token(lemma_="hola", pos_="hola"),
+        Token(lemma_="hola", pos_="hola"),
+        Token(lemma_="chao", pos_="chao"),
+        Token(lemma_="hola", pos_="hola"),
+        Token(lemma_="perro", pos_="perro"),
+        Token(lemma_="hola", pos_="hola"),
+    ]
+    yield doc
 
 
 def test_type_token_ratio():
@@ -22,36 +42,51 @@ def test_one_side_lexical_diversity_mtld():
     )
 
 
-def test_lexical_diversity_mtld():
+def test_lexical_diversity_mtld(test_doc):
     """Test lexical_diversity_mtld."""
-    Token = namedtuple("Token", "lemma_ pos_")
-    doc = [
-        Token("hola", "hola"),
-        Token("hola", "hola"),
-        Token("chao", "chao"),
-        Token("hola", "hola"),
-        Token("perro", "perro"),
-        Token("hola", "hola"),
-    ]
-    assert ttr.lexical_diversity_mtld(doc, ttr_segment=1) == 3
+    assert ttr.lexical_diversity_mtld(test_doc, ttr_segment=1) == 3
 
 
-def test_yule_k():
+def test_yule_k(test_doc):
     """Test yule_k."""
-    Token = namedtuple("Token", "lemma_ pos_")
-    doc = [
-        Token("hola", "hola"),
-        Token("hola", "hola"),
-        Token("chao", "chao"),
-        Token("hola", "hola"),
-        Token("perro", "perro"),
-        Token("hola", "hola"),
-    ]
-
-    n = len(doc)
+    n = len(test_doc)
     rs = {
         1: 2,
         4: 1,
     }
     expected_k = 1e-4 * sum(r ** 2 * vr - n for r, vr in rs.items()) / n ** 2
-    assert ttr.yule_k(doc) == expected_k
+    assert ttr.yule_k(test_doc) == expected_k
+
+
+def test_d_estimate():
+    """Test d_estimate."""
+    text = (
+        "El espermatozoide y el ovocito son células son muy diferentes entre "
+        "sí, y poseen propiedades estructurales que van adquiriendo mediante "
+        "el proceso de gametogénesis. La gametogénesis masculina se denomina "
+        "espermatogénesis y la femenina, ovogénesis. Ocurre al interior de "
+        "los testículos, en unas estructuras llamadas túbulos seminíferos Se "
+        "inicia en la pubertad y, en condiciones normales, se mantiene "
+        "durante toda la vida de los hombres. A continuación, revisaremos sus "
+        "etapas. Durante el desarrollo embrionario, las células germinales "
+        "primordiales se multiplican, dando lugar a espermatogonias. Años más "
+        "tarde, en la pubertad, algunas espermatogonias proliferan, aumentan "
+        "de tamaño y se diferencian en espermatocitos primarios o "
+        "espermatocitos I. Luego, los espermatocitos I pasan por un proceso "
+        "que consta de dos divisiones celulares. La primera división, origina "
+        "los espermatocitos secundarios o espermatocitos II; estas células "
+        "experimentan la segunda división formando las espermátidas. Cada una "
+        "de ellas tiene la mitad del material genético de la especie. "
+        "Finalmente, las espermátidas experimentan cambios morfológicos que "
[…]
