### 2021-04-15 0baeeb9
Add D-estimate and update refs. (# 29)
