+++ docs/api_reference/ttr.bib
@@ -1,0 +1,14 @@
+@article{richards2000measuring,
+  title={Measuring vocabulary richness in teenage learners of French},
+  author={Richards, Brian and Malvern, David},
+  year={2000},
+  publisher={Education-line},
+  journal={British Educational Research Association Annual Conference}
+}
+
+@book{yule2014statistical,
+  title={The statistical study of literary vocabulary},
+  author={Yule, C Udny},
+  year={2014},
+  publisher={Cambridge University Press}
+}
