+++ docs/api_reference/ttr.rst
@@ -5,3 +5,5 @@ Type Token Ratios
 
 .. automodule:: TRUNAJOD.ttr
     :members:
+
+.. bibliography:: ttr.bib
