### 2021-04-15 a03a090
add type hints
