### 2021-04-15 6041a39
Implement Guiraud’s Index
