### 2021-04-15 9b86aae
Implement Guiraud’s Index
