### 2021-04-15 dpalmasan/TRUNAJOD2.0#50 @mbdemoraes
State of the field
In terms of state of the field, the paper include a lot of discussions around text measurements. But how about competitor libraries? TRUNAJOD is the only library/package (including other programming languages) that serves this purpose? If it isn't the only one,  how does it works compared to those implementations?

cc https://github.com/openjournals/joss-reviews/issues/3153
