### 2021-04-15 dpalmasan/TRUNAJOD2.0#50 @dpalmasan
Okay, it seems I need to update the `paper.md` with such discussion. Thanks for pointing out! I will let you know once I fix this!
