### 2021-04-15 dpalmasan/TRUNAJOD2.0#51 @mbdemoraes
Include additional examples on the documentation
In the documentation, there is only one example of usage. Since there is a lot of features, the authors should include at least two or three more examples, using other features to improve the understanding of the proposed package. Currently, the remaining functionality is described only at an API level, which seems insufficient.

cc https://github.com/openjournals/joss-reviews/issues/3153
