### 2021-04-15 #3153 @mbdemoraes
@dpalmasan great job so far! I believe these should be the last issues:

- https://github.com/dpalmasan/TRUNAJOD2.0/issues/50
- https://github.com/dpalmasan/TRUNAJOD2.0/issues/51
