### 2021-04-15 dpalmasan/TRUNAJOD2.0#51 @dpalmasan
Got it. I will include a more detailed example of usage. I will let you know when this is done!
