### 2021-04-15 0afdba5
Update README and paper per JOSS review.
