+++ paper.bib
@@ -74,6 +74,14 @@
   }
 }
 
+@inproceedings{smith2016cohere,
+  title={Cohere: A toolkit for local coherence},
+  author={Smith, Karin Sim and Aziz, Wilker and Specia, Lucia},
+  booktitle={Proceedings of the Tenth International Conference on Language Resources and Evaluation (LREC'16)},
+  pages={4111--4114},
+  year={2016}
+}
+
 @article{guasch2016spanish,
   title={Spanish norms for affective and lexico-semantic variables for 1,400 words},
   author={Guasch, Marc and Ferr{\'e}, Pilar and Fraga, Isabel},
@@ -88,6 +96,21 @@
   URL = { 
           https://doi.org/10.3758/s13428-015-0684-y
       
+  },
+}
+
+@article{crossley2019tool,
+  title={The Tool for the Automatic Analysis of Cohesion 2.0: Integrating semantic similarity and text overlap},
+  author={Crossley, Scott A and Kyle, Kristopher and Dascalu, Mihai},
+  journal={Behavior research methods},
+  volume={51},
+  number={1},
+  pages={14--27},
+  year={2019},
+  publisher={Springer},
+  doi = {10.3758/s13428-018-1142-4},
+  URL = {
+    https://doi.org/10.3758/s13428-018-1142-4
   },
 }
 
