### 2021-04-15 284a32e
Update README and paper per JOSS review.
