+++ paper.md
@@ -37,14 +37,15 @@ We present `TRUNAJOD`, a text complexity analysis tool that includes a wide vari
 2. Having an open-source code, so any researcher in the linguistics field could contribute to it
 3. Making easy to build applications and tools that rely upon text complexity assessment
 
+Other tools aim to make it easy for the public to get coherence and cohesion metrics. One such tool is TAACO [@crossley2019tool], which is written in python and can be freely downloaded. The problem with TAACO is that it is a desktop application, which encloses the code. This makes it impossible to contribute with modifications or new features, as it is a closed system. Moreover, it does not implement other relevant features to assess cohesion and coherence of discourse, for example, entity grid-based features. One open-source project that aims to do this is [@smith2016cohere], which is written in a mix of Java and Python. However, it seems it is not actively maintained and it does not implement other measurements that could be used by other researchers. On the other hand, most of the tools only support English languages and do not provide support for a plethora of metrics available in a comprehensible API. `TRUNAJOD` aims to be different, in the sense that we do not present a closed system, but we present an open-source project, trying to follow the best development patterns for the Python languages, and we also rely on `spaCy` lib, to not only support one language but multiple languages for coherence and cohesion tasks, making possible to improve `TRUNAJOD` performance if `spaCy` performance improves, which promotes collaboration.
+
 Moreover, `TRUNAJOD` is not only taking the state of the art measurements for text complexity assessment, but also comes bundled with new sets of predictors for this task, and thus `TRUNAJOD` ́s contributions are:
 
 * Fixes periphrasis of texts, because many NLP tools have issues dealing with periphrasis. In this release, this only applies to Spanish.
 * Adds heuristics for measurements based on clause count. TRUNAJOD provides a new algorithm for clause segmentation.
 * TRUNAJOD provides several approximations to narrativity in these new clause segmentation-dependent indices.
 
-Text complexity assessment is a natural language processing task that can be applied for multiple problems, such as automatic summarization, automatic essay scoring, automatic summary evaluation, intelligent tutoring systems, and so on. Text complexity is usually related to the readability of a text, which is dependent on several intrinsic 
-properties of it, mainly cohesion and coherence.
+Text complexity assessment is a natural language processing task that can be applied for multiple problems, such as automatic summarization, automatic essay scoring, automatic summary evaluation, intelligent tutoring systems, and so on. Text complexity is usually related to the readability of a text, which is dependent on several intrinsic properties of it, mainly cohesion and coherence.
 
 Automatic coherence evaluation is an open problem, and there have been several studies addressing it. On one side, text coherence assessment has been related to how sentences connect either semantically, 
 or by co-referencing noun phrases. In the semantic view of coherence, Latent Semantic Analysis (LSA) [@foltz1998measurement] has been widely used because of its simplicity. In essence, sentences are represented as vectors, and the coherence of the text is computed as the average sentence similarity, using similarity vector measurements (such as cosine distance). This approach has drawbacks, such as that sentence ordering does not matter. For solving this issue, other methodologies have been proposed, based on discourse theory, in particular the centering theory. One such approach is entity grids [@barzilay2008modeling] and entity graphs [@guinaudeau2013graph] that treat coherence as to how are entities taking different roles between sentences and how are they connected in the text. `TRUNAJOD` comes with all these models implemented, and thus `TRUNAJOD` can compute coherence based on sentence similarities using word vectors, and also `TRUNAJOD` has an API for dealing with entity grids and entity graphs, to extract such measurements.
