### 2021-04-15 dpalmasan/TRUNAJOD2.0#50 @dpalmasan
@mbdemoraes I updated the `paper.md` file to provide more details for the comparisons with other libs. Let me know if anything is missing. I will re-generate the paper on JOSS site.
