### 2021-04-15 dpalmasan/TRUNAJOD2.0#51 @dpalmasan
@mbdemoraes I included a code snippet showing how this can be used on research related to text analysis. I also provided a real application demo (as a youtube video) built on `TRUNAJOD`. Let me know if I need to improve anything else. Thanks for the review BTW!
