### 2021-04-15 dpalmasan/TRUNAJOD2.0#51 @mbdemoraes
@dpalmasan I've checked the new examples, thank you for including them. By the way, very good video!

Since the modification addressed the suggestion, I'm going to close the issue now.
