### 2021-04-15 dpalmasan/TRUNAJOD2.0#50 @mbdemoraes
@dpalmasan well done! Everything is fine now.
