### 2021-04-16 dpalmasan/TRUNAJOD2.0#52 @apiad
Complete / expand contribution guidelines
Hey :wave: ! This issue is part of the JOSS review (openjournals/joss-reviews#3153) for this project.

First of all, kudos on the project! As a Spanish-speaker NLP researcher myself, I honestly cannot commend you enough regarding the quality and breadth of this work. We need this kind of research in our community and I thank you on behalf of all my colleague researchers as well.

Checking over the CONTRIBUTING guidelines, it seems there are some TBD sections that could be completed, not necessarily in great detail, but at least to a point where they're useful. If you feel there is a section that is unnecessary you can just remove it (e.g., how to create a pull-request is a general Github-level issue that I don't necessarily think your guidelines should spend too much effort in explaining).

My suggestion is to review the CONTRIBUTING guidelines and either add some content in those TBD sections that you consider relevant or reformat the document to remove them, thus making the guidelines complete and welcoming to new contributors. Please let me know if you feel this is a big issue or if you want to further discuss it.
