### 2021-04-16 #3153 @apiad
@dpalmasan I just checked all the functional claims and documentation. I opened a (hopefully easy) issue to improve the contribution guidelines just a little bit.

I'm now reviewing the paper itself, and I have some suggestions. I'll finish reviewing it and open a single issue with all of them. Most of them are gonna are optional suggestions that I think can improve the quality of the writing, nothing too complex, mostly rewriting some long sentences and clarifying some points. I hope these suggestions help!

Great work, BTW. I already said it elsewhere but as a Spanish-speaker NLP researcher, I'm very happy to see this kind of high-quality work in our native language.
