+++ CONTRIBUTING.md
@@ -6,17 +6,32 @@ The following is a set of guidelines for contributing to `TRUNAJOD`. These are m
 
 #### Table Of Contents
 
-This section is TBD.
+[How Can I Contribute?](#how-can-i-contribute)
+  * [Reporting Bugs](#reporting-bugs)
+  * [Suggesting Enhancements](#suggesting-enhancements)
+  * [Your First Code Contribution](#your-first-code-contribution)
+  * [Pull Requests](#pull-requests)
+
+[Styleguides](#styleguides)
+  * [Git Commit Messages](#git-commit-messages)
+  * [Python Styleguide](#python-styleguide)
+  * [Documentation Styleguide](#documentation-styleguide)
 
 ## How Can I Contribute?
 
 ### Reporting Bugs
 
-This section is TBD.
+If you encounter a bug or something in the functionality does not seem right, feel free to file an issue in the Github issue tracker. Ideally, provide an isolated example on how to reproduce the issue. The minimal information to help us assess the bug/issue:
+
+* Clear description of the bug
+* Current behavior and expected behavior
+* Ideally a minimal example on how to reproduce it. We know that it is not always straightforward doing this, so this is optional.
 
 ### Suggesting Enhancements
 
-This section is TBD.
+Feel free to suggest any enhancement. Note that if you'd like a certain new index to be implemented, please provide in the description a reference paper or article, so we understand better what the enhancement is about and how can we make sure it is correctly implemented. 
+
+We accept any kind of enhancements, so if you feel that usability, API, documentation can be improved, feel free to submit an issue.
 
 ### Your First Code Contribution
 
@@ -27,7 +42,21 @@ Unsure where to begin contributing to `TRUNAJOD`? You can start by looking throu
 
 ### Pull Requests
 
-TBD
+Feel free to submit a pull request if you have a proposal for a feature or bugfix. However, make sure that the changes description are clear and you provide unit tests for the new code. Moreover, make sure that all the worflows in the CI (Github actions) pass, as we'd prefer not having regressions on the test cases. Finally, make sure to add citations if you add an algorithm. This will involve updating the `.bib` file for the updated module and making sure that you add the citation in the docstring of the new function. For example:
+
+```python
+def d_estimate(
+    doc: Doc, min_range: int = 35, max_range: int = 50, trials: int = 5
+) -> float:
+    r"""Compute D measurement for lexical diversity.
+
+    The measurement is based in :cite:`richards2000measuring`. We pick ``n``
+    numbers of tokens, varying ``N`` from ``min_range`` up to ``max_range``.
+    For each ``n`` we do the following:
+
+    # Continues...
+    """
+```
 
 ## Styleguides
 
@@ -52,21 +81,13 @@ Therefore, before committing, we recommend installing [pre-commit](https://pre-c
 * `pip install pre-commit`
 * `pre-commit install`
 
-Now, before committing, `pre-commit` hooks will be run, and you can ensure the checks are passed before the commit is created.
+Now, before committing, `pre-commit` hooks will be run, and you can ensure the checks are passed before the commit is created. Finally, make sure that you use typehints for the new functions you define, as we are planning to add `mypy` checks to the workflows.
 
 ### Documentation Styleguide
 
 We use restructured text format, this is important as the documentation is hosted in [Read the Docs](https://readthedocs.org/). In particular, `TRUNAJOD` docs can be found in [https://trunajod20.readthedocs.io/en/latest/](https://trunajod20.readthedocs.io/en/latest/).
 
-#### Example
-
-To be added.
-
-## Additional Notes
-
-### Issue and Pull Request Labels
-
[…]
