### 2021-04-16 dpalmasan/TRUNAJOD2.0#52 @dpalmasan
Hello @apiad , thank you for your kind words and also thank you for the detailed review. I totally agree with you with the `CONTRIBUTING.md` updates. I have already updated the guidelines and provided descriptions on the `TBD` placeholders. Let me know if I missed anything from your suggestions.
