### 2021-04-16 #3153 @apiad
@danielskatz what's in your opinion the best UX for a writing/grammar review? Should I open an issue and just say "where you say ... in line X" or is there a better way to use the Github UI, maybe opening a PR with suggested changes, etc.?

What I would ideally prefer is to go over `paper.md` and just add comments on specific lines, so the PR workflow seems perfect for this, but to open a PR I have to propose some changes and *then* add comments on the PR. Is there a better/suggested way? Thanks in advance.
