### 2021-04-16 dpalmasan/TRUNAJOD2.0#52 @apiad
Whoa, so fast! It seems complete now in my opinion.
