### 2021-04-16 14c7ad6
Add TRUNAJOD logo
