### 2021-04-16 181b68c
Add TRUNAJOD logo
