### 2021-04-16 ff0c73d
Update trunajod logo size.
