+++ src/TRUNAJOD/semantic_measures.py
@@ -6,10 +6,9 @@ synonym overlap measurement requires knowledge from a word onthology, and
 semantic measurements require word vectors (word embeddings) obtained from
 CORPUS semantics.
 """
-from spacy.tokens import Doc
 
 
-def avg_w2v_semantic_similarity(docs: Doc, N: int) -> float:
+def avg_w2v_semantic_similarity(docs, N):
     """Compute average semantic similarity between adjacent sentences.
 
     This is using word2vec :cite:`mikolov2013word2vec` model based on SPACY
@@ -44,7 +43,7 @@ def avg_w2v_semantic_similarity(docs: Doc, N: int) -> float:
     return avg_sim / float(N - 1)
 
 
-def get_synsets(lemma: str, synset_dict: dict) -> str:
+def get_synsets(lemma, synset_dict):
     """Return synonym set given a word lemma.
 
     The function requires that the synset_dict is passed into it. In our case
@@ -62,7 +61,7 @@ def get_synsets(lemma: str, synset_dict: dict) -> str:
     return synset_dict.get(lemma, {lemma})
 
 
-def overlap(lemma_list_group: list, synset_dict: dict) -> float:
+def overlap(lemma_list_group, synset_dict):
     """Compute average overlap in a text.
 
     Computes semantic synset overlap (synonyms), based on a lemma list group
