+++ src/TRUNAJOD/lexico_semantic_norms.py
@@ -13,7 +13,6 @@ compute the following lexico-semantic variables:
 We provide two downloadable models of these variables, which come from
 :cite:`duchon2013espal` and :cite:`guasch2016spanish`.
 """
-from spacy.tokens import Doc
 from TRUNAJOD.lexicosemantic_norms_espal import LEXICOSEMANTIC_ESPAL
 from TRUNAJOD.lexicosemantic_norms_espal import LSNorm
 from TRUNAJOD.utils import lemmatize
@@ -29,12 +28,7 @@ class LexicoSemanticNorm(object):
     :cite:`guasch2016spanish`.
     """
 
-    def __init__(
-        self,
-        doc: Doc,
-        lexico_semantic_norm_dict: dict,
-        lemmatizer: dict = None,
-    ):
+    def __init__(self, doc, lexico_semantic_norm_dict, lemmatizer=None):
         """Initialize lexico semantic norm object.
 
         Calculate average over number of tokens given a text.
@@ -108,7 +102,7 @@ class LexicoSemanticNorm(object):
             self.__context_avilability /= count
             self.__familiarity /= count
 
-    def get_arousal(self) -> float:
+    def get_arousal(self):
         """Get arousal.
 
         :return: Average arousal.
@@ -116,7 +110,7 @@ class LexicoSemanticNorm(object):
         """
         return self.__arousal
 
-    def get_concreteness(self) -> float:
+    def get_concreteness(self):
         """Get concreteness.
 
         :return: Average concreteness.
@@ -124,7 +118,7 @@ class LexicoSemanticNorm(object):
         """
         return self.__concreteness
 
-    def get_context_availability(self) -> float:
+    def get_context_availability(self):
         """Get context_availability.
 
         :return: Average context_availability.
@@ -132,7 +126,7 @@ class LexicoSemanticNorm(object):
         """
         return self.__context_avilability
 
-    def get_familiarity(self) -> float:
+    def get_familiarity(self):
         """Get familiarity.
 
         :return: Average familiarity.
@@ -140,7 +134,7 @@ class LexicoSemanticNorm(object):
         """
         return self.__familiarity
 
-    def get_imageability(self) -> float:
+    def get_imageability(self):
         """Get imageability.
 
         :return: Average imageability.
@@ -148,7 +142,7 @@ class LexicoSemanticNorm(object):
         """
         return self.__imageability
 
-    def get_valence(self) -> float:
+    def get_valence(self):
         """Get valence.
 
         :return: Average valence.
@@ -157,7 +151,7 @@ class LexicoSemanticNorm(object):
         return self.__valence
 
 
-def get_conc_imag_familiarity(doc: Doc) -> [float, float, float]:
+def get_conc_imag_familiarity(doc):
     """Get lexico-semantic variables.
 
     Computes three lexico-semantic variables: Concreteness, Imageability and
