### 2021-04-18 faa2e93
First iteration of model install mechanism.
