### 2021-04-19 4551455
Add load metadata utility (#16)
