+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-We present `TRUNAJOD`, a text complexity analysis tool that includes a wide variety of linguistics measurements that can be extracted for texts as an approximate of readability, coherence, and cohesion. The features that `TRUNAJOD` can extract from the text are based on the literature and can be separated into the following categories: Discourse markers, emotions, entity grid-based measurements, givenness, lexical-semantic norms, semantic measures, surface proxies, etc. In this first version of `TRUNAJOD`, we support mainly the Spanish language, but several features support any language that has proper natural language processing POS tagging and dependency parsing capabilities. Finally, we show how TRUNAJOD could be used in applied research.
+We present `TRUNAJOD`, a text complexity analysis tool that includes a wide variety of linguistics measurements that can be extracted from texts as an approximation for readability, coherence, and cohesion. The features that `TRUNAJOD` can extract from the text are based on the literature and can be separated into the following categories: Discourse markers, emotions, entity grid-based measurements, givenness, lexical-semantic norms, semantic measures, surface proxies, etc. In this first version of `TRUNAJOD`, we support mainly the Spanish language, but several features support any language that has proper natural language processing POS tagging and dependency parsing capabilities. Finally, we show how TRUNAJOD could be used in applied research.
 
 # Statement of need
 
@@ -37,29 +37,27 @@ We present `TRUNAJOD`, a text complexity analysis tool that includes a wide vari
 2. Having an open-source code, so any researcher in the linguistics field could contribute to it
 3. Making easy to build applications and tools that rely upon text complexity assessment
 
-Other tools aim to make it easy for the public to get coherence and cohesion metrics. One such tool is TAACO [@crossley2019tool], which is written in python and can be freely downloaded. The problem with TAACO is that it is a desktop application, which encloses the code. This makes it impossible to contribute with modifications or new features, as it is a closed system. Moreover, it does not implement other relevant features to assess cohesion and coherence of discourse, for example, entity grid-based features. One open-source project that aims to do this is [@smith2016cohere], which is written in a mix of Java and Python. However, it seems it is not actively maintained and it does not implement other measurements that could be used by other researchers. On the other hand, most of the tools only support English languages and do not provide support for a plethora of metrics available in a comprehensible API. `TRUNAJOD` aims to be different, in the sense that we do not present a closed system, but we present an open-source project, trying to follow the best development patterns for the Python languages, and we also rely on `spaCy` lib, to not only support one language but multiple languages for coherence and cohesion tasks, making possible to improve `TRUNAJOD` performance if `spaCy` performance improves, which promotes collaboration.
[…]
