### 2021-04-20 #3153 @apiad
Hey @dpalmasan, just submitted a small PR with what I think could be minor fixes to the paper. Please forgive me for taking the freedom to directly suggest text changes, but I think is easier/faster than listing them here. Let me know if you wish to review anything, and of course, as the author, you have the final word on your paper. Hope my suggestions help a bit :smile:!
