+++ paper.md
@@ -37,7 +37,7 @@ We present `TRUNAJOD`, a text complexity analysis tool that includes a wide vari
 2. Having an open-source code, so any researcher in the linguistics field could contribute to it
 3. Making easy to build applications and tools that rely upon text complexity assessment
 
-Other tools aim to make it easy for the public to get coherence and cohesion metrics. One such tool is TAACO [@crossley2019tool], which is written in Python and can be freely downloaded. The problem with TAACO is that it is a desktop application, which encloses the code. This makes it impossible to contribute with modifications or new features, as it is a closed system. Moreover, it does not implement other relevant features to assess cohesion and coherence of discourse, for example, entity grid-based features. One open-source project with this purpose is <NAME?> [@smith2016cohere], which is written in a mixture of Java and Python. However, it does not seem actively maintained and it does not implement other measurements that could be used by other researchers. On the other hand, most of the tools only support English languages and do not provide support for a plethora of metrics available in a comprehensible API. `TRUNAJOD` aims to be different, in the sense that we do not present a closed system, but we present an open-source project, trying to follow the best development patterns for the Python language. Furthermore, we rely on `spaCy`, to not only support one language but multiple languages for coherence and cohesion tasks, enabling `TRUNAJOD` to improve its performance if `spaCy` does, which promotes collaboration.
+Other tools aim to make it easy for the public to get coherence and cohesion metrics. One such tool is TAACO [@crossley2019tool], which is written in Python and can be freely downloaded. The problem with TAACO is that it is a desktop application, which encloses the code. This makes it impossible to contribute with modifications or new features, as it is a closed system. Moreover, it does not implement other relevant features to assess cohesion and coherence of discourse, for example, entity grid-based features. One open-source project with this purpose is `Cohere` framework [@smith2016cohere], which is written in a mixture of Java and Python. However, it does not seem actively maintained and it does not implement other measurements that could be used by other researchers. On the other hand, most of the tools only support English languages and do not provide support for a plethora of metrics available in a comprehensible API. `TRUNAJOD` aims to be different, in the sense that we do not present a closed system, but we present an open-source project, trying to follow the best development patterns for the Python language. Furthermore, we rely on `spaCy`, to not only support one language but multiple languages for coherence and cohesion tasks, enabling `TRUNAJOD` to improve its performance if `spaCy` does, which promotes collaboration.
 
 Moreover, `TRUNAJOD` not only implements state of the art measurements for text complexity assessment, but also bundles new sets of predictors for this task. In this sense, `TRUNAJOD` ́s contributions are:
 
