### 2021-04-21 #3153 @danielskatz
👋 @dpalmasan - I've suggested a bunch of small changes to the paper and bib in https://github.com/dpalmasan/TRUNAJOD2.0/pull/57
