### 2021-04-21 #3153 @dpalmasan
Hello @danielskatz, I already merged your changes, thanks! Here are the items of the checklist:

- Tagged release with the reviewed software `v.0.1.2`: https://github.com/dpalmasan/TRUNAJOD2.0/releases/tag/v0.1.2
- Zenodo software archival: https://zenodo.org/record/4707403#.YIAa5RNKh70
- Metadata title and authors match the `paper.md`
- The archived version DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4707403.svg)](https://doi.org/10.5281/zenodo.4707403)

Please let me know if further changes are needed
