### 2021-04-21 #3153 @danielskatz
Congratulations to @dpalmasan (Diego Palma) and co-authors on your publication!!

And thanks to @mbdemoraes and @apiad for reviewing!

As the editor, this was a very quick and smooth process that worked well as everyone was quite collaborative and responsive - thank again to all of you!!
